Use double instead of int for force and displacement

+++ include/Filter_Alg.h
@@ -2,24 +2,24 @@
 #define _FILTERALG_H
 
 typedef struct _SumPara{
-	float sum_X;
-	float sum_Y;
-	float sum_Xpow2;
-	float sum_Xpow3;
-	float sum_Xpow4;
-	float sum_XY;
-	float sum_Xpow2Y;
+	double sum_X;
+	double sum_Y;
+	double sum_Xpow2;
+	double sum_Xpow3;
+	double sum_Xpow4;
+	double sum_XY;
+	double sum_Xpow2Y;
 }SumPara;
 
 
 typedef struct _FuncPara{
-	float a;
-	float b;
-	float c;
+	double a;
+	double b;
+	double c;
 }FuncPara;
 
 
-void CalFuncPara(int force, int displace, int count);
+void CalFuncPara(double force, double displace, int count);
 
 extern FuncPara funcParaDisplacement;
 extern FuncPara funcParaForce;
